+++ src/app/profile/actions/change-password-action.ts
@@ -70,8 +70,8 @@ export const changePasswordAction = authActionClient()
                 })
                 .where(eq(users.id, userId));
 
-            revalidatePath('/user');
-            revalidatePath(`/user/${user.id}`);
+            revalidatePath('/users');
+            revalidatePath(`/users/${user.id}`);
             revalidatePath('/profile');
 
             return { id: user.id };
