+++ src/app/users/[id]/update-user-form.tsx
@@ -1,7 +1,7 @@
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
-import { GetUserResponse } from '@/app/user/data/get-user';
+import { GetUserResponse } from '@/app/users/data/get-user';
 import { TextField } from '@/components/form/text-field';
 import { DateField } from '@/components/form/date-field';
 import { SwitchField } from '@/components/form/switch-field';
