+++ src/app/users/create/page.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from 'next';
-import { CreateUser } from '@/app/user/create/create-user';
+import { CreateUser } from '@/app/users/create/create-user';
 import { verifySession } from '@/lib/dal';
 import { isUserAdministrator } from '@/lib/definitions';
